social-app-backend/django_core/requirements.txt
Django==5.0.1
celery==5.3.4
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2

psycopg2-binary==2.9.9
django-environ==0.11.2
python-decouple==3.8

django-cors-headers==4.3.1
django-filter==23.5

redis==5.0.1
django-redis==5.4.0

gunicorn==21.2.0
channels==4.0.0
daphne==4.0.0
whitenoise==6.6.0
requests==2.31.0
pillow==10.2.0
boto3==1.34.10
